requirements.txt
streamlit
requests
beautifulsoup4
faiss-cpu
scikit-learn
google-generativeai
tqdm
lxml
langchain-community